Extract execution tags into own component
