+++ packages/frontend/editor-ui/src/components/executions/workflow/WorkflowExecutionsPreview.test.ts
@@ -137,92 +137,6 @@ describe('WorkflowExecutionsPreview.vue', () => {
 		expect(getByTestId('stop-execution')).toBeDisabled();
 	});
 
-	it('displays annotated tags', async () => {
-		const executionWithTags = {
-			...executionData,
-			annotation: {
-				tags: [
-					{ id: 'tag1', name: 'Test Tag 1' },
-					{ id: 'tag2', name: 'Test Tag 2' },
-				],
-				vote: 'up' as AnnotationVote,
-			},
-		};
-
-		// Set up the test with the execution having tags
-		const pinia = createTestingPinia({
-			initialState: {
-				[STORES.SETTINGS]: {
-					settings: {
-						enterprise: {
-							[EnterpriseEditionFeature.AdvancedExecutionFilters]: true,
-						},
-					},
-				},
-				[STORES.EXECUTIONS]: {
-					activeExecution: executionWithTags,
-				},
-			},
-		});
-
-		const { getByTestId, queryByTestId } = renderComponent({
-			props: { execution: executionWithTags },
-			pinia,
-		});
-
-		await nextTick();
-
-		// Should show annotation tags container
-		expect(getByTestId('annotation-tags-container')).toBeInTheDocument();
-		// Should show existing tags (not the dropdown)
-		expect(getByTestId('execution-annotation-tags')).toBeInTheDocument();
-		// Should not show the dropdown when tags exist
-		expect(queryByTestId('workflow-tags-dropdown')).not.toBeInTheDocument();
-	});
-
-	it('should be able to add a tag', async () => {
-		const executionWithoutTags = {
-			...executionData,
-			annotation: { tags: [], vote: 'up' as AnnotationVote },
-		};
-
-		// Set up the test with the execution having no tags
-		const pinia = createTestingPinia({
-			initialState: {
-				[STORES.SETTINGS]: {
-					settings: {
-						enterprise: {
-							[EnterpriseEditionFeature.AdvancedExecutionFilters]: true,
-						},
-					},
-				},
-				[STORES.EXECUTIONS]: {
-					activeExecution: executionWithoutTags,
-				},
-			},
-		});
-
-		const { getByTestId, queryByTestId } = renderComponent({
-			props: { execution: executionData },
-			pinia,
-		});
-
-		await nextTick();
-
-		// Should show the add tag button when no tags exist
-		const addTagButton = getByTestId('new-tag-link');
-		expect(addTagButton).toBeInTheDocument();
-
-		// Should not show the dropdown initially
-		expect(queryByTestId('workflow-tags-dropdown')).not.toBeInTheDocument();
-
-		// Click the add tag button
-		await userEvent.click(addTagButton);
-
-		// Should show the dropdown after clicking
-		expect(getByTestId('workflow-tags-dropdown')).toBeInTheDocument();
-	});
-
 	it('should display vote buttons when annotation is enabled', async () => {
 		// Set up the test with annotation enabled
 		const pinia = createTestingPinia({
@@ -254,45 +168,6 @@ describe('WorkflowExecutionsPreview.vue', () => {
 		// Should contain two button elements (thumbs up and thumbs down)
 		const buttons = voteButtons.querySelectorAll('button');
 		expect(buttons).toHaveLength(2);
-	});
-
-	it('should handle vote button clicks', async () => {
-		// Set up the test with annotation enabled
-		const pinia = createTestingPinia({
-			initialState: {
-				[STORES.SETTINGS]: {
-					settings: {
-						enterprise: {
-							[EnterpriseEditionFeature.AdvancedExecutionFilters]: true,
-						},
-					},
-				},
-				[STORES.EXECUTIONS]: {
-					activeExecution: executionData,
-				},
-			},
-		});
-
-		const { getByTestId } = renderComponent({
-			props: { execution: executionData },
-			pinia,
-		});
-
-		await nextTick();
-
-		const voteButtons = getByTestId('execution-preview-vote-buttons');
-		const buttons = voteButtons.querySelectorAll('button');
-
-		expect(buttons).toHaveLength(2);
-
-		// Click the first button (thumbs up)
-		await userEvent.click(buttons[0]);
-		// Click the second button (thumbs down)
-		await userEvent.click(buttons[1]);
-
-		// The buttons should still be present after clicking
-		expect(buttons[0]).toBeInTheDocument();
-		expect(buttons[1]).toBeInTheDocument();
 	});
 
 	it('should show active vote state', async () => {
@@ -403,82 +278,6 @@ describe('WorkflowExecutionsPreview.vue', () => {
 		// Should not show badge when no custom data
 		const badge = ellipsisButton.querySelector('.badge');
 		expect(badge).not.toBeInTheDocument();
-	});
-
-	it('should show add tag button when no tags exist', async () => {
-		const executionWithoutTags = {
-			...executionData,
-			annotation: {
-				tags: [],
-				vote: 'up' as AnnotationVote,
-			},
-		};
-
-		// Set up the test with no tags
-		const pinia = createTestingPinia({
-			initialState: {
-				[STORES.SETTINGS]: {
-					settings: {
-						enterprise: {
-							[EnterpriseEditionFeature.AdvancedExecutionFilters]: true,
-						},
-					},
-				},
-				[STORES.EXECUTIONS]: {
-					activeExecution: executionWithoutTags,
-				},
-			},
-		});
-
-		const { getByTestId, queryByTestId } = renderComponent({
-			props: { execution: executionWithoutTags },
-			pinia,
-		});
-
-		await nextTick();
-
-		// Should show the add tag button
-		expect(getByTestId('new-tag-link')).toBeInTheDocument();
-		// Should not show existing tags container
-		expect(queryByTestId('execution-annotation-tags')).not.toBeInTheDocument();
-	});
-
-	it('should show existing tags with add button when tags exist', async () => {
-		const executionWithTags = {
-			...executionData,
-			annotation: {
-				tags: [{ id: 'tag1', name: 'Test Tag 1' }],
-				vote: 'up' as AnnotationVote,
-			},
-		};
-
-		// Set up the test with existing tags
-		const pinia = createTestingPinia({
-			initialState: {
-				[STORES.SETTINGS]: {
-					settings: {
-						enterprise: {
-							[EnterpriseEditionFeature.AdvancedExecutionFilters]: true,
-						},
-					},
-				},
-				[STORES.EXECUTIONS]: {
-					activeExecution: executionWithTags,
-				},
-			},
-		});
-
-		const { getByTestId } = renderComponent({
-			props: { execution: executionWithTags },
-			pinia,
-		});
-
-		await nextTick();
-
-		// Should show existing tags container
-		expect(getByTestId('execution-annotation-tags')).toBeInTheDocument();
-		// Should also show add tag button for adding more tags
-		expect(getByTestId('new-tag-link')).toBeInTheDocument();
 	});
 
 	it('should not show vote buttons when annotation is disabled', async () => {
